requirements.txt
requests
Pillow
requests
Pillow
python-dotenv